requirements.txt
ffmpeg-python
yt-dlp 
openai-whisper 
googletrans 
pyaudio 
numpy
# brew install portaudio  